wip(pdm): add hook
