--- bases/polylith/pdm_hooks/__init__.py
@@ -1,3 +1,0 @@
-from polylith.pdm_hooks import core
-
-__all__ = ["core"]
