+++ bases/polylith/pdm_hooks/core.py
@@ -1,7 +1,11 @@
+from pathlib import Path
+
 from polylith import pdm
 
 
 def pdm_build_initialize(context):
     context.ensure_build_dir()
 
-    pdm.hooks.bricks.build_initialize(context.config.data)
+    build_dir = Path(context.build_dir)
+
+    pdm.hooks.bricks.build_initialize(context.config.data, build_dir)
